Numbers on Motor Insurance set to Int

+++ app/src/main/java/model/AvailableMotorInsuranceModel.java
@@ -1,7 +1,8 @@
 package model;
 
 public interface AvailableMotorInsuranceModel {
-       double getPrice();
+       int getPrice();
+
        double getRadioValue();
        double getWindScreenValue();
